requirements.txt
# Core dependencies
pyTelegramBotAPI==4.14.0
selenium==4.15.2
webdriver-manager==4.0.1
faker==20.1.0
python-dotenv==1.0.0

# Additional utilities
requests==2.31.0
pillow==10.1.0
python-magic==0.4.27
cryptography==41.0.7

# Logging and monitoring
loguru==0.7.2
psutil==5.9.6

# Cloud deployment
gunicorn==21.2.0
flask==3.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
